[Update] Index design improvement
