+++ src/pages/components/Layout.tsx
@@ -12,7 +12,7 @@ export default function Layout({ children }: LayoutProps){
     <>
         <Navbar />
             {children}
-        <Footer />
+        {/* <Footer /> */}
     </>
     );
 }
