+++ src/styles/cat.scss
@@ -1,28 +1,63 @@
 .catconteiner {
     display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
+    flex-wrap: wrap;
+    
+    //flex-direction: row;
+    // align-items: center;
+    // justify-content: center;
     //margin: 0 20px 0 -20px;
+    gap: 10px;
     width: 100%;
-    height: 100%;
+    height: 80%;
     //background-color: #f5f5f5;
     padding: 20px;
     .catleft {
         display: flex;
+        flex: 1;
+        flex-direction: column-reverse;
+        align-items: flex-end;
+        justify-content: center;
+        background: url('https://img.freepik.com/free-vector/cute-shiba-inu-dog-sushi-cartoon-icon-illustration_138676-2413.jpg?w=900&t=st=1683218556~exp=1683219156~hmac=648bda97bcf590efb58096f91f1992a14cc582ed69c4f5f620a1ab8a7ea274aa');
+        background-size: cover;
+        width: 50%;
+        height: 50%;
+        padding: 20px;
+        border-radius: 10px;
+        //border: red solid 1px;
+    }
+    .catbuttom{
+        display: flex;
         flex-direction: column;
+        align-items: flex-start;
+        justify-content: flex-start;
+        background: url('https://img.freepik.com/premium-vector/three-little-kittens-cartoon-illustration_241806-629.jpg?w=1380');
+        background-size: cover;
+        width: 100%;
+        height: 50%;
+        padding: 20px;
+        border-radius: 10px;
+        //border: red solid 1px;
+    }
+    .catright {
+        display: flex;
+        flex: 1;
+        flex-direction: column;
+        background: url('https://img.freepik.com/free-vector/hand-drawn-kawaii-illustration_52683-85164.jpg?w=1060&t=st=1683218436~exp=1683219036~hmac=fe90bce9b6fcedd1b9f2fd5ff5202c213d4f0ffdd132753aa60706160d1427b9');
+        background-size: cover;
         align-items: center;
         justify-content: center;
+        border-radius: 10px;
         width: 50%;
-        height: 100%;
+        height: 50%;
         padding: 20px;
-        border: red solid 1px;
-        .catname {
-            font-size: 2rem;
-            font-weight: bold;
-            color: #333;
-            margin-bottom: 20px;
+        //border: red solid 1px;
+        .catimg {
+            width: 100%;
+            height: 100%;
+            //border-radius: 10px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
         }
+
         .catinfo {
             display: flex;
             flex-direction: column;
@@ -59,34 +94,24 @@
     }
 }
 
-.catright {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 50%;
-    height: 100%;
-    padding: 20px;
-    border: red solid 1px;
-    .catimg {
-        width: 100%;
-        height: 100%;
-        border-radius: 10px;
-        box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-    }
+
+.catname {
+    font-size: 3rem;
+    font-weight: bold;
+    //color: #00ffaa;
+    margin-bottom: 20px;
 }
 
 .catbutton {
-    //background-color: transparent;
-    //background: url('https://media1.giphy.com/media/jTf83ZnIn7A1dKRqvQ/giphy.gif?cid=ecf05e47ti6k8eslzgdxz5h9bu99d6c5s18wtg6oskz2jzeq&ep=v1_gifs_search&rid=giphy.gif&ct=g');
+    background-color: transparent;
     border: #333 solid 1px;
     transition: all 0.3s ease-in-out;
     &:hover {
-        background: url('https://media1.giphy.com/media/jTf83ZnIn7A1dKRqvQ/giphy.gif?cid=ecf05e47ti6k8eslzgdxz5h9bu99d6c5s18wtg6oskz2jzeq&ep=v1_gifs_search&rid=giphy.gif&ct=g');
         background-size: cover;
         transform: scale(1.1);
-        background-color: #333;
-        color: #fff;
+        background-color: #29292974;
+        color: #ffffff83; 
+        box-shadow: 5px 5px 0px 2px rgba(0, 0, 0, 0.5);
     }
 }
 
@@ -111,8 +136,6 @@
     margin: auto;
     border: red solid 1px;
     max-width: 1200px;
-    //flex: 2;
-    
     .catitle {
         text-align: center;
         font-size: 2rem;
@@ -142,7 +165,6 @@
     &:hover {
         transform: scale(1.1);
     }
-
   }
 
       
@@ -164,3 +186,4 @@
     //width: calc(33.33% - 4px);
     margin: 2px;
   }
+
