+++ src/styles/globals.scss
@@ -34,8 +34,6 @@ body {
   --custom-border-radius: 5px;
   --custom-spacing: 5px;
 
-  padding: 0;
-  margin: 0;
   //font-family: var(--custom-font-family);
   background-color: var(--custom-bg-color);
 }
