+++ src/styles/navbar.scss
@@ -1,24 +1,35 @@
 nav {
     display: flex;
-    justify-content: space-between;
+    justify-content: center;
     align-items: center;
-    
+
     padding: 0 20px;
-    position: relative;
+    //position: relative;
     top: 0;
     width: 100%;
-    z-index: 999;
+    //z-index: 999;
+
+    .navbar.fixed {
+        //position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        background-color: #fff;
+        box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+      }
 
     .navcont {
         display: flex;
         flex-direction: column;
         align-items: center;
         justify-content: center;
+        position: relative;
+        border-radius: 0 10px 10px 0;
         margin: 0 100px 0 100px;
         width: 100%;
         height: 100%;
         background-color: #ecf3f812;
-        border-radius: 10px;
+        //border-radius: 10px;
         border-bottom: 1px solid #e5e5e5;
         box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
     }
